requirements.txt
opencv-python
numpy
telepot
twilio
RPi.GPIO
picamera
imutils
python-dotenv